+++ Changelog.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file. This file a
 
 * Updating Cargo dependencies
 
+### Removed:
+* Removed flags for filename entry and part_number entry (as they're default and required every time)
+
 ### Added
 
 * Changelog!
+* Added export shortages function
+* Added export shortages command